Add ISO link to ISO download dialog
This changes the discovery modal behavior - after the download is ready
the modal does not close and provides an ISO URL to be coppied and a button
to trigger a download via browser.
