+++ src/components/clusterConfiguration/DiscoveryImageSummary.tsx
@@ -1,0 +1,73 @@
+import React from 'react';
+import { saveAs } from 'file-saver';
+import {
+  Button,
+  ButtonVariant,
+  ClipboardCopy,
+  clipboardCopyFunc,
+  EmptyState,
+  EmptyStateIcon,
+  EmptyStateVariant,
+  Title,
+  ModalBoxBody,
+  ModalBoxFooter,
+  TextContent,
+} from '@patternfly/react-core';
+import { global_success_color_100 as successColor } from '@patternfly/react-tokens';
+import { CheckCircleIcon } from '@patternfly/react-icons';
+import { Cluster } from '../../api/types';
+import { getClusterDownloadsImageUrl } from '../../api/clusters';
+import { DetailList, DetailItem } from '../ui/DetailList';
+
+type DiscoveryImageSummaryProps = {
+  clusterId: Cluster['id'];
+  onClose: () => void;
+  onReset: () => void;
+};
+
+const DiscoveryImageSummary: React.FC<DiscoveryImageSummaryProps> = ({
+  clusterId,
+  onClose,
+  onReset,
+}) => {
+  const isoPath = getClusterDownloadsImageUrl(clusterId);
+  const isoUrl = `${window.location.origin}${isoPath}`;
+  const downloadIso = () => saveAs(isoPath);
+  return (
+    <>
+      <ModalBoxBody>
+        <EmptyState variant={EmptyStateVariant.small}>
+          <EmptyStateIcon icon={CheckCircleIcon} color={successColor.value} />
+          <Title headingLevel="h4" size="lg">
+            Discovery ISO is ready to download
+          </Title>
+        </EmptyState>
+        <TextContent>
+          <DetailList>
+            <DetailItem
+              title="Discovery ISO URL"
+              value={
+                <ClipboardCopy isReadOnly onCopy={(event) => clipboardCopyFunc(event, isoUrl)}>
+                  {isoUrl}
+                </ClipboardCopy>
+              }
+            />
+          </DetailList>
+        </TextContent>
+      </ModalBoxBody>
+      <ModalBoxFooter>
+        <Button variant={ButtonVariant.primary} onClick={downloadIso}>
+          Download Discovery ISO
+        </Button>
+        <Button variant={ButtonVariant.secondary} onClick={onClose}>
+          Close
+        </Button>
+        <Button variant={ButtonVariant.link} onClick={onReset}>
+          Edit ISO Configuration
+        </Button>
+      </ModalBoxFooter>
+    </>
+  );
+};
+
+export default DiscoveryImageSummary;
